requirements.txt
streamlit
pillow
opencv
numpy
pandas
tensorflow
scikit-learn
